requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-spectacular==0.17.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
